+++ src/pipe_utils/values.py
@@ -24,6 +24,7 @@ __all__ = [
     "rclamp",
     "is_none",
     "is_not_none",
+    "raise_",
     "add_by",
     "sub_by",
     "mul_by",
@@ -40,6 +41,8 @@ __all__ = [
     "ge",
     "it_is",
     "it_is_not",
+    "is_",
+    "is_not",
     "bit_and",
     "bit_or",
     "bit_xor",
@@ -92,18 +95,6 @@ def is_congruent(a: int, n: int) -> Callable[[int], bool]:
     return lambda item: (item - a) % n == 0
 
 
-@deprecated(version="0.2.0", reason="Use :const:`pipe_utils.pipe.it` instead")
-def is_non_negative(item: SupportsGE) -> bool:
-    """Returns True if the given item is greater than or equal to 0"""
-    return item >= 0
-
-
-@deprecated(version="0.2.0", reason="Use :const:`pipe_utils.pipe.it` instead")
-def is_non_positive(item: SupportsLE) -> bool:
-    """Returns True if the given item is less than or equal to 0"""
-    return item <= 0
-
-
 def clamp(lower: T, upper: T) -> Callable[[T], T]:
     """
     Returns a callable that clamps a value between lower and upper.
@@ -136,6 +127,39 @@ def is_none(item: Any) -> bool:
 def is_not_none(item: Any) -> bool:
     """Returns True if the given item is not None"""
     return item is not None
+
+
+def is_(obj: Any) -> Callable:
+    """Returns a callable that takes a parameter x and returns x is n"""
+    return lambda item: operator.is_(item, obj)
+
+
+def is_not(obj: Any) -> Callable:
+    """Returns a callable that takes a parameter x and returns x is not n"""
+    return lambda item: operator.is_not(item, obj)
+
+
+def raise_(exception: E | EType, *, from_: E = None):
+    """
+    Raise expression. Raises exceptions. If ``from`` is given, will raise the
+    exception from that.
+    """
+    if from_ is None:
+        raise exception
+    else:
+        raise exception from from_
+
+
+@deprecated(version="0.2.0", reason="Use :const:`pipe_utils.pipe.it` instead")
+def is_non_negative(item: SupportsGE) -> bool:
+    """Returns True if the given item is greater than or equal to 0"""
+    return item >= 0
+
+
+@deprecated(version="0.2.0", reason="Use :const:`pipe_utils.pipe.it` instead")
+def is_non_positive(item: SupportsLE) -> bool:
+    """Returns True if the given item is less than or equal to 0"""
+    return item <= 0
 
 
 @deprecated(version="0.2.0", reason="Use :const:`pipe_utils.pipe.it` instead")
@@ -222,14 +246,17 @@ def ge(other: Any) -> Callable:
     return lambda item: operator.ge(item, other)
 
 
+@deprecated(version="0.2.1", reason="Use :func:`pipe_utils.values.is_` instead")
 def it_is(obj: Any) -> Callable:
     """Returns a callable that takes a parameter x and returns x is n"""
     return lambda item: operator.is_(item, obj)
 
-
+@deprecated(version="0.2.1",
+            reason="Use :func:`pipe_utils.values.is_not` instead")
 def it_is_not(obj: Any) -> Callable:
     """Returns a callable that takes a parameter x and returns x is not n"""
     return lambda item: operator.is_not(item, obj)
+
 
 
 @deprecated(version="0.2.0", reason="Use :const:`pipe_utils.pipe.it` instead")